requirements.txt
albumentations==0.4.3
opencv-python==4.1.2.30
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.3.5
pytorch-lightning-bolts
omegaconf==2.0.6
test-tube>=0.7.5
streamlit>=0.73.1